+++ test/specs/openapi/test_parameters.py
@@ -1,7 +1,14 @@
 import pytest
 
 from schemathesis.specs.openapi.definitions import OPENAPI_30_VALIDATOR, SWAGGER_20_VALIDATOR
-from schemathesis.specs.openapi.parameters import OpenAPI20Body, OpenAPI20Parameter, OpenAPI30Parameter, OpenAPIBody
+from schemathesis.specs.openapi.parameters import (
+    OpenAPI20Body,
+    OpenAPI20CompositeBody,
+    OpenAPI20Parameter,
+    OpenAPI30Body,
+    OpenAPI30Parameter,
+    OpenAPIBody,
+)
 
 
 @pytest.mark.parametrize(
@@ -78,3 +85,50 @@ def test_examples(request, cls, schema, expected):
     kwargs = {"media_type": "text/plain"} if issubclass(cls, OpenAPIBody) else {}
 
     assert cls(schema, **kwargs).example == expected
+
+
+DESCRIPTION = "Foo"
+
+
+@pytest.mark.parametrize(
+    "cls, kwargs, expected",
+    (
+        (
+            OpenAPI20Parameter,
+            {"definition": {"description": DESCRIPTION, "in": "query", "name": "foo", "type": "string"}},
+            DESCRIPTION,
+        ),
+        (
+            OpenAPI30Parameter,
+            {"definition": {"description": DESCRIPTION, "in": "query", "name": "foo", "schema": {"type": "string"}}},
+            DESCRIPTION,
+        ),
+        (
+            OpenAPI20Body,
+            {
+                "definition": {"description": DESCRIPTION, "in": "body", "name": "foo", "schema": {"type": "string"}},
+                "media_type": "application/json",
+            },
+            DESCRIPTION,
+        ),
+        (
+            OpenAPI30Body,
+            {
+                "definition": {"schema": {"type": "string"}},
+                "description": DESCRIPTION,
+                "media_type": "application/json",
+            },
+            DESCRIPTION,
+        ),
+        (
+            OpenAPI20CompositeBody,
+            {
+                "definition": [OpenAPI20Parameter({"in": "formData", "name": "foo", "type": "string"})],
+                "media_type": "application/x-www-form-urlencoded",
+            },
+            None,
+        ),
+    ),
+)
+def test_description(cls, kwargs, expected):
+    assert cls(**kwargs).description == expected