+++ src/schemathesis/cli/output/default.py
@@ -286,6 +286,9 @@ def handle_service_integration(context: ExecutionContext) -> None:
         status = click.style(event.name, fg=color, bold=True)
         click.echo(f"{title}: {status}\r", nl=False)
         click.echo()
+        if isinstance(event, service.Success):
+            report_title = click.style("Report", bold=True)
+            click.echo(f"{report_title}: {event.short_url}")
         if isinstance(event, service.Error):
             click.echo()
             if context.show_errors_tracebacks: