+++ src/schemathesis/service/models.py
@@ -1,0 +1,7 @@
+import attr
+
+
+@attr.s(slots=True)
+class TestJob:
+    job_id: str = attr.ib()
+    short_url: str = attr.ib()