+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 - New ``before_init_operation`` hook.
 - **INTERNAL**. ``description`` attribute for all parsed parameters inside ``APIOperation``.
+- Timeouts when loading external schema components or external examples.
 
 **Changed**